Change total costs to int

+++ src/pharmCounting.py
@@ -79,7 +79,7 @@ class pharmRecord(object):
                 continue
 
             try:
-                float(line_array[-1]) # check if the last element is a valid float. If not, skip this line.
+                int(line_array[-1]) # check if the last element is a valid float. If not, skip this line.
             except ValueError:
                 if self.logging:
                     self.log(status='WARNING', 
@@ -90,13 +90,13 @@ class pharmRecord(object):
 
             patient_name = ','.join(line_array[1:-2]) # patients' last and first name are joined using ','
             drug_name = line_array[-2] # assume the second last element is the drug name
-            drug_cost = float(line_array[-1]) # assume the last element is the drug cost. 
+            drug_cost = int(line_array[-1]) # assume the last element is the drug cost. 
 
             if drug_name not in self.records: # make a new entry for a drug if not in the dictionary
                 self.records[drug_name] = {'patients' : set(), 'costs' : 0}
 
             self.records[drug_name]['patients'].add(patient_name) # use hashset to remove name duplication
-            self.records[drug_name]['costs'] += float(drug_cost) # sum all the costs for a drug
+            self.records[drug_name]['costs'] += drug_cost # sum all the costs for a drug
             
         if self.logging:
             self.log(status='NORMAL',
